test: relocate dbMock helper under src/test and adjust test imports
